Add generic JoinEntity class and fix DTOs

+++ nxplx-backend/NxPlx.Integrations.TMDB/Models/DetailsBase.cs
@@ -10,11 +10,11 @@ namespace NxPlx.Integrations.TMDBApi.Models
         public string homepage { get; set; }
         public string original_language { get; set; }
         public string overview { get; set; }
-        public double popularity { get; set; }
+        public float popularity { get; set; }
         public string poster_path { get; set; }
         public List<ProductionCompany> production_companies { get; set; }
         public string status { get; set; }
-        public double vote_average { get; set; }
+        public float vote_average { get; set; }
         public int vote_count { get; set; }
     }
 }
