+++ nxplx-backend/NxPlx.Models/Details/Search/ResultBase.cs
@@ -5,7 +5,7 @@ namespace NxPlx.Models.Details.Search
     public abstract class ResultBase
     {
         public int Id { get; set; }
-        public double Popularity { get; set; }
+        public float Popularity { get; set; }
         public string OriginalLanguage { get; set; }
         public List<int> GenreIds { get; set; }
         public string Overview { get; set; }
